requirements.txt
# === Core Django & Utilities ===
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
psycopg[binary]==3.1.18  # PostgreSQL support

# === Authentication & Security ===
argon2-cffi==23.1.0
bcrypt==5.0.0
django-allauth==65.12.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8

# === REST & API ===
drf-yasg==1.21.7  # Swagger/OpenAPI docs
djangorestframework==3.16.1

# === Real-time & Channels ===
channels==4.0.0
channels-redis==4.2.0

# === Cloud Storage (Optional) ===
boto3==1.34.0
cloudinary==1.38.0
django-storages==1.14.2

# === Observability & Monitoring ===
coverage==7.11.0
sentry-sdk==2.0.1

# === CLI & Dev Tools ===
click==8.3.0
colorama==0.4.6
packaging==25.0
python-dotenv==1.0.1

# === Flask (Optional microservices or admin tools) ===
Flask==3.1.2
Flask-Bcrypt==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
blinker==1.9.0

# === Server ===
gunicorn==23.0.0
